refactor: ProductImage 컴포넌트 공통화
