+++ app/src/main/java/nextstep/shoppingcart/ui/screen/ShoppingCartScreen.kt
@@ -2,8 +2,19 @@ package nextstep.shoppingcart.ui.screen
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun ShoppingCartScreen(modifier: Modifier = Modifier) {
+fun ShoppingCartScreen(
+    modifier: Modifier = Modifier,
+) {
 
+}
+
+
+
+@Preview(showBackground = true)
+@Composable
+fun ShoppingCartScreenPreview() {
+    ShoppingCartScreen()
 }
