+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/ProductImage.kt
@@ -1,0 +1,30 @@
+package nextstep.shoppingcart.ui.screen.component
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.tooling.preview.Preview
+import coil3.compose.AsyncImage
+
+@Composable
+fun ProductImage(
+    imageUrl: String,
+    contentDescription: String,
+    modifier: Modifier = Modifier,
+) {
+    AsyncImage(
+        model = imageUrl,
+        contentDescription = contentDescription,
+        contentScale = ContentScale.Crop,
+        modifier = modifier
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun ProductImagePreview() {
+    ProductImage(
+        imageUrl = "https://picsum.photos/200/300",
+        contentDescription = "Android Robot"
+    )
+}
