+++ app/src/main/java/nextstep/shoppingcart/ui/screen/component/ProductItem.kt
@@ -7,11 +7,9 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
-import coil3.compose.AsyncImage
 import nextstep.shoppingcart.ui.utils.formatter.DefaultMoneyFormatter
 import nextstep.shoppingcart.ui.utils.formatter.MoneyFormatter
 
@@ -27,10 +25,9 @@ fun ProductItem(
         modifier = modifier,
         horizontalAlignment = Alignment.Start,
     ) {
-        AsyncImage(
-            model = imageUrl,
+        ProductImage(
+            imageUrl = imageUrl,
             contentDescription = "Product Image",
-            contentScale = ContentScale.Crop,
             modifier = Modifier
                 .fillMaxWidth()
                 .aspectRatio(78f / 79f)
